Move task SQL into `TaskRepository`
This moves CAS1 specific code out of a generic repository. Also the queries are specifically related to tasks.
