+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/jpa/entity/TaskEntity.kt
@@ -199,6 +199,141 @@ interface TaskRepository : JpaRepository<Task, UUID> {
     completed: Boolean,
     pageable: Pageable?,
   ): Page<Task>
+
+  @Query(
+    """
+    SELECT u.*,
+     RANK() OVER (
+      ORDER BY
+        (
+          SELECT COUNT(1)
+            FROM assessments a
+            WHERE a.allocated_to_user_id = u.id
+            AND 
+              (
+                a.submitted_at IS NULL
+                OR (
+                  a.submitted_at IS NOT NULL
+                    AND a.created_at BETWEEN 
+                      (CURRENT_TIMESTAMP - interval '1 week') AND
+                  		CURRENT_TIMESTAMP
+                )
+            )
+        ) ASC
+     ) as score
+    FROM "users" u
+    WHERE u.id IN (:userIds)
+    ORDER BY score ASC
+    LIMIT 1
+    """,
+    nativeQuery = true,
+  )
+  fun findUserWithLeastAssessmentsPendingOrCompletedInLastWeek(userIds: List<UUID>): UserEntity?
+
+  @Query(
+    """
+    SELECT u.*,
+    RANK() OVER (
+      ORDER BY
+        (
+          SELECT COUNT(1)
+            FROM placement_applications pa
+            WHERE pa.allocated_to_user_id = u.id
+            AND 
+              (
+                pa.submitted_at IS NULL
+                OR (
+                  pa.submitted_at IS NOT NULL
+                    AND pa.created_at BETWEEN 
+                      (CURRENT_TIMESTAMP - interval '1 week') AND
+                  		CURRENT_TIMESTAMP
+                )
+            )
+        ) ASC
+     ) as score
+    FROM "users" u
+    WHERE u.id IN (:userIds)
+    ORDER BY score ASC 
+    LIMIT 1
+    """,
+    nativeQuery = true,
+  )
+  fun findUserWithLeastPlacementApplicationsPendingOrCompletedInLastWeek(userIds: List<UUID>): UserEntity?
+
+  @Query(
+    """
+    SELECT
+      CAST(u.id as TEXT) as userId,
+      (
+        SELECT
+          count(*)
+        from
+          assessments a
+        where
+          a.allocated_to_user_id = u.id
+          and a.reallocated_at is null
+          and a.submitted_at is null
+          and a.is_withdrawn != true
+      ) as pendingAssessments,
+      (
+        SELECT
+          count(*)
+        from
+          assessments a
+        where
+          a.allocated_to_user_id = u.id
+          and a.reallocated_at is null
+          and a.submitted_at > current_date - interval '7' day
+      ) as completedAssessmentsInTheLastSevenDays,
+      (
+        SELECT
+          count(*)
+        from
+          assessments a
+        where
+          a.allocated_to_user_id = u.id
+          and a.reallocated_at is null
+          and a.submitted_at > current_date - interval '30' day
+      ) as completedAssessmentsInTheLastThirtyDays,
+      (
+        SELECT
+          count(*)
+        from
+          placement_applications placement_application
+        where
+          placement_application.allocated_to_user_id = u.id
+          and placement_application.reallocated_at is null
+          and placement_application.submitted_at is null
+          and placement_application.decision is null
+      ) as pendingPlacementApplications,
+      (
+        SELECT
+          count(*)
+        from
+          placement_applications placement_application
+        where
+          placement_application.allocated_to_user_id = u.id
+          and placement_application.reallocated_at is null
+          and placement_application.submitted_at > current_date - interval '7' day
+      ) as completedPlacementApplicationsInTheLastSevenDays,
+      (
+        SELECT
+          count(*)
+        from
+          placement_applications placement_application
+        where
+          placement_application.allocated_to_user_id = u.id
+          and placement_application.reallocated_at is null
+          and placement_application.submitted_at > current_date - interval '30' day
+      ) as completedPlacementApplicationsInTheLastThirtyDays
+    FROM
+      users u
+    WHERE
+      u.id IN (:userIds)
+    """,
+    nativeQuery = true,
+  )
+  fun findWorkloadForUserIds(userIds: List<UUID>): List<UserWorkload>
 }
 
 @Entity
