+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/unit/service/cas1/Cas1TaskServiceTest.kt
@@ -32,7 +32,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.Task
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.TaskEntityType
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.TaskRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserEntity
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.PaginationMetadata
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.CasResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.AssessmentService
@@ -57,7 +56,6 @@ class Cas1TaskServiceTest {
   private val taskRepositoryMock = mockk<TaskRepository>()
   private val assessmentRepositoryMock = mockk<AssessmentRepository>()
   private val placementApplicationRepositoryMock = mockk<PlacementApplicationRepository>()
-  private val userRepositoryMock = mockk<UserRepository>()
 
   private val cas1TaskService = Cas1TaskService(
     assessmentServiceMock,
@@ -68,7 +66,6 @@ class Cas1TaskServiceTest {
     taskRepositoryMock,
     assessmentRepositoryMock,
     placementApplicationRepositoryMock,
-    userRepositoryMock,
   )
 
   private val requestUser = UserEntityFactory().withDefaults().produce()
